Minor changes in the way of detecting the links
Cause a problem with FF and the way it handles the header detection
for CORS request, we are not able to get the content length of the
file, so we don't know if it's 0 or the size of the book.

Cause project gutenberg is returning 200 for urls that doesn't exists
at all :P bad boy.